refactoring to pull out code for commutators
