+++ src/openfermion/utils/_dual_basis_trotter_error_test.py
@@ -19,229 +19,6 @@ from openfermion.utils._dual_basis_trotter_error import *
 from openfermion.utils import Grid
 
 
-class DoubleCommutatorTest(unittest.TestCase):
-
-    def test_double_commutator_no_intersection_with_union_of_second_two(self):
-        com = double_commutator(FermionOperator('4^ 3^ 6 5'),
-                                FermionOperator('2^ 1 0'),
-                                FermionOperator('0^'))
-        self.assertTrue(com.isclose(FermionOperator.zero()))
-
-    def test_double_commutator_more_info_not_hopping(self):
-        com = double_commutator(
-            FermionOperator('3^ 2'),
-            FermionOperator('2^ 3') + FermionOperator('3^ 2'),
-            FermionOperator('4^ 2^ 4 2'), indices2=set([2, 3]),
-            indices3=set([2, 4]), is_hopping_operator2=True,
-            is_hopping_operator3=False)
-        self.assertTrue(com.isclose(FermionOperator('4^ 2^ 4 2') -
-                                    FermionOperator('4^ 3^ 4 3')))
-
-    def test_double_commtator_more_info_both_hopping(self):
-        com = double_commutator(
-            FermionOperator('4^ 3^ 4 3'),
-            FermionOperator('1^ 2', 2.1) + FermionOperator('2^ 1', 2.1),
-            FermionOperator('1^ 3', -1.3) + FermionOperator('3^ 1', -1.3),
-            indices2=set([1, 2]), indices3=set([1, 3]),
-            is_hopping_operator2=True, is_hopping_operator3=True)
-        self.assertTrue(com.isclose(FermionOperator('4^ 3^ 4 2', 2.73) +
-                                    FermionOperator('4^ 2^ 4 3', 2.73)))
-
-
-class TriviallyDoubleCommutesDualBasisUsingTermInfoTest(unittest.TestCase):
-
-    def test_number_operators_trivially_commute(self):
-        self.assertTrue(trivially_double_commutes_dual_basis_using_term_info(
-            indices_alpha=set([1, 2]), indices_beta=set([3, 4]),
-            indices_alpha_prime=set([2, 3]),
-            is_hopping_operator_alpha=False, is_hopping_operator_beta=False,
-            is_hopping_operator_alpha_prime=False, jellium_only=True))
-
-    def test_left_hopping_operator_no_trivial_commutation(self):
-        self.assertFalse(trivially_double_commutes_dual_basis_using_term_info(
-            indices_alpha=set([1, 2]), indices_beta=set([3, 4]),
-            indices_alpha_prime=set([2, 3]),
-            is_hopping_operator_alpha=True, is_hopping_operator_beta=True,
-            is_hopping_operator_alpha_prime=False, jellium_only=True))
-
-    def test_right_hopping_operator_no_trivial_commutation(self):
-        self.assertFalse(trivially_double_commutes_dual_basis_using_term_info(
-            indices_alpha=set([1, 2]), indices_beta=set([3, 4]),
-            indices_alpha_prime=set([2, 3]),
-            is_hopping_operator_alpha=True, is_hopping_operator_beta=False,
-            is_hopping_operator_alpha_prime=True, jellium_only=True))
-
-    def test_alpha_is_hopping_operator_others_number_trivial_commutation(self):
-        self.assertTrue(trivially_double_commutes_dual_basis_using_term_info(
-            indices_alpha=set([1, 2]), indices_beta=set([3, 4]),
-            indices_alpha_prime=set([2, 3]),
-            is_hopping_operator_alpha=True, is_hopping_operator_beta=False,
-            is_hopping_operator_alpha_prime=False, jellium_only=True))
-
-    def test_no_intersection_in_first_commutator_trivially_commutes(self):
-        self.assertTrue(trivially_double_commutes_dual_basis_using_term_info(
-            indices_alpha=set([1, 2]), indices_beta=set([3, 4]),
-            indices_alpha_prime=set([1, 2]),
-            is_hopping_operator_alpha=True, is_hopping_operator_beta=True,
-            is_hopping_operator_alpha_prime=False, jellium_only=True))
-
-    def test_double_intersection_in_first_commutator_trivially_commutes(self):
-        self.assertTrue(trivially_double_commutes_dual_basis_using_term_info(
-            indices_alpha=set([3, 2]), indices_beta=set([3, 4]),
-            indices_alpha_prime=set([4, 3]),
-            is_hopping_operator_alpha=True, is_hopping_operator_beta=True,
-            is_hopping_operator_alpha_prime=False, jellium_only=True))
-
-    def test_single_intersection_in_first_commutator_nontrivial(self):
-        self.assertFalse(trivially_double_commutes_dual_basis_using_term_info(
-            indices_alpha=set([3, 2]), indices_beta=set([3, 4]),
-            indices_alpha_prime=set([4, 5]),
-            is_hopping_operator_alpha=False, is_hopping_operator_beta=True,
-            is_hopping_operator_alpha_prime=False, jellium_only=True))
-
-    def test_no_intersection_between_first_and_other_terms_is_trivial(self):
-        self.assertTrue(trivially_double_commutes_dual_basis_using_term_info(
-            indices_alpha=set([3, 2]), indices_beta=set([1, 4]),
-            indices_alpha_prime=set([4, 5]),
-            is_hopping_operator_alpha=False, is_hopping_operator_beta=True,
-            is_hopping_operator_alpha_prime=False, jellium_only=True))
-
-
-class TriviallyCommutesDualBasisTest(unittest.TestCase):
-
-    def test_trivially_commutes_no_intersection(self):
-        self.assertTrue(trivially_commutes_dual_basis(
-            FermionOperator('3^ 2^ 3 2'), FermionOperator('4^ 1')))
-
-    def test_no_trivial_commute_with_intersection(self):
-        self.assertFalse(trivially_commutes_dual_basis(
-            FermionOperator('2^ 1'), FermionOperator('5^ 2^ 5 2')))
-
-    def test_trivially_commutes_both_single_number_operators(self):
-        self.assertTrue(trivially_commutes_dual_basis(
-            FermionOperator('3^ 3'), FermionOperator('3^ 3')))
-
-    def test_trivially_commutes_nonintersecting_single_number_operators(self):
-        self.assertTrue(trivially_commutes_dual_basis(
-            FermionOperator('2^ 2'), FermionOperator('3^ 3')))
-
-    def test_trivially_commutes_both_double_number_operators(self):
-        self.assertTrue(trivially_commutes_dual_basis(
-            FermionOperator('3^ 2^ 3 2'), FermionOperator('3^ 1^ 3 1')))
-
-    def test_trivially_commutes_one_double_number_operators(self):
-        self.assertTrue(trivially_commutes_dual_basis(
-            FermionOperator('3^ 2^ 3 2'), FermionOperator('3^ 3')))
-
-    def test_no_trivial_commute_right_hopping_operator(self):
-        self.assertFalse(trivially_commutes_dual_basis(
-            FermionOperator('3^ 1^ 3 1'), FermionOperator('3^ 2')))
-
-    def test_no_trivial_commute_left_hopping_operator(self):
-        self.assertFalse(trivially_commutes_dual_basis(
-            FermionOperator('3^ 2'), FermionOperator('3^ 3')))
-
-    def test_trivially_commutes_both_hopping_create_same_mode(self):
-        self.assertTrue(trivially_commutes_dual_basis(
-            FermionOperator('3^ 2'), FermionOperator('3^ 1')))
-
-    def test_trivially_commutes_both_hopping_annihilate_same_mode(self):
-        self.assertTrue(trivially_commutes_dual_basis(
-            FermionOperator('4^ 1'), FermionOperator('3^ 1')))
-
-    def test_trivially_commutes_both_hopping_and_number_on_same_modes(self):
-        self.assertTrue(trivially_commutes_dual_basis(
-            FermionOperator('4^ 1'), FermionOperator('4^ 1^ 4 1')))
-
-
-class TriviallyDoubleCommutesDualBasisTest(unittest.TestCase):
-
-    def test_trivially_double_commutes_no_intersection(self):
-        self.assertTrue(trivially_double_commutes_dual_basis(
-            FermionOperator('3^ 4'),
-            FermionOperator('3^ 2^ 3 2'), FermionOperator('4^ 1')))
-
-    def test_no_trivial_double_commute_with_intersection(self):
-        self.assertFalse(trivially_double_commutes_dual_basis(
-            FermionOperator('4^ 2'),
-            FermionOperator('2^ 1'), FermionOperator('5^ 2^ 5 2')))
-
-    def test_trivially_double_commutes_both_single_number_operators(self):
-        self.assertTrue(trivially_double_commutes_dual_basis(
-            FermionOperator('4^ 3'),
-            FermionOperator('3^ 3'), FermionOperator('3^ 3')))
-
-    def test_trivially_double_commutes_nonintersecting_single_number_ops(self):
-        self.assertTrue(trivially_double_commutes_dual_basis(
-            FermionOperator('3^ 2'),
-            FermionOperator('2^ 2'), FermionOperator('3^ 3')))
-
-    def test_trivially_double_commutes_both_double_number_operators(self):
-        self.assertTrue(trivially_double_commutes_dual_basis(
-            FermionOperator('4^ 3'),
-            FermionOperator('3^ 2^ 3 2'), FermionOperator('3^ 1^ 3 1')))
-
-    def test_trivially_double_commutes_one_double_number_operators(self):
-        self.assertTrue(trivially_double_commutes_dual_basis(
-            FermionOperator('4^ 3'),
-            FermionOperator('3^ 2^ 3 2'), FermionOperator('3^ 3')))
-
-    def test_no_trivial_double_commute_right_hopping_operator(self):
-        self.assertFalse(trivially_double_commutes_dual_basis(
-            FermionOperator('4^ 3'),
-            FermionOperator('3^ 1^ 3 1'), FermionOperator('3^ 2')))
-
-    def test_no_trivial_double_commute_left_hopping_operator(self):
-        self.assertFalse(trivially_double_commutes_dual_basis(
-            FermionOperator('4^ 3'),
-            FermionOperator('3^ 2'), FermionOperator('3^ 3')))
-
-    def test_trivially_double_commutes_both_hopping_create_same_mode(self):
-        self.assertTrue(trivially_double_commutes_dual_basis(
-            FermionOperator('3^ 3'),
-            FermionOperator('3^ 2'), FermionOperator('3^ 1')))
-
-    def test_trivially_double_commutes_both_hopping_annihilate_same_mode(self):
-        self.assertTrue(trivially_double_commutes_dual_basis(
-            FermionOperator('1^ 1'),
-            FermionOperator('4^ 1'), FermionOperator('3^ 1')))
-
-    def test_trivially_double_commutes_hopping_and_number_on_same_modes(self):
-        self.assertTrue(trivially_double_commutes_dual_basis(
-            FermionOperator('4^ 3'),
-            FermionOperator('4^ 1'), FermionOperator('4^ 1^ 4 1')))
-
-    def test_trivially_double_commutes_no_intersection_a_with_bc(self):
-        self.assertTrue(trivially_double_commutes_dual_basis(
-            FermionOperator('5^ 2'),
-            FermionOperator('3^ 1'), FermionOperator('4^ 1^ 4 1')))
-
-    def test_trivially_double_commutes_double_create_in_a_and_b(self):
-        self.assertTrue(trivially_double_commutes_dual_basis(
-            FermionOperator('5^ 2'),
-            FermionOperator('3^ 1'), FermionOperator('4^ 1^ 4 1')))
-
-    def test_trivially_double_commutes_double_annihilate_in_a_and_c(self):
-        self.assertTrue(trivially_double_commutes_dual_basis(
-            FermionOperator('5^ 2'),
-            FermionOperator('3^ 1'), FermionOperator('4^ 1^ 4 1')))
-
-    def test_no_trivial_double_commute_double_annihilate_with_create(self):
-        self.assertFalse(trivially_double_commutes_dual_basis(
-            FermionOperator('5^ 2'),
-            FermionOperator('2^ 1'), FermionOperator('4^ 2')))
-
-    def test_trivially_double_commutes_excess_create(self):
-        self.assertTrue(trivially_double_commutes_dual_basis(
-            FermionOperator('5^ 2'),
-            FermionOperator('5^ 5'), FermionOperator('5^ 1')))
-
-    def test_trivially_double_commutes_excess_annihilate(self):
-        self.assertTrue(trivially_double_commutes_dual_basis(
-            FermionOperator('5^ 2'),
-            FermionOperator('3^ 2'), FermionOperator('2^ 2')))
-
-
 class ErrorOperatorTest(unittest.TestCase):
 
     def test_error_operator(self):
@@ -485,7 +262,3 @@ class OrderedDualBasisTermsNoInfoTest(unittest.TestCase):
         hamiltonian = jellium_model(grid, spinless=True, plane_wave=False)
         hamiltonian = normal_ordered(hamiltonian)
         self.assertTrue(terms_total.isclose(hamiltonian))
-
-
-if __name__ == '__main__':
-    unittest.main()
