+++ src/openfermion/utils/_operator_utils.py
@@ -70,29 +70,6 @@ def count_qubits(operator):
         raise TypeError('Operator of invalid type.')
 
 
-def commutator(operator_a, operator_b):
-    """Compute the commutator of two operators.
-
-    Args:
-        operator_a, operator_b: Operators in commutator. Any operators
-            are accepted so long as implicit subtraction and multiplication are
-            supported; e.g. QubitOperators, FermionOperators or Scipy sparse
-            matrices. 2D Numpy arrays are also supported.
-
-    Raises:
-        TypeError: operator_a and operator_b are not of the same type.
-    """
-    if type(operator_a) != type(operator_b):
-        raise TypeError('operator_a and operator_b are not of the same type.')
-    if isinstance(operator_a, numpy.ndarray):
-        result = operator_a.dot(operator_b)
-        result -= operator_b.dot(operator_a)
-    else:
-        result = operator_a * operator_b
-        result -= operator_b * operator_a
-    return result
-
-
 def eigenspectrum(operator):
     """Compute the eigenspectrum of an operator.
